main/requirements.txt
pandas==1.3.5
protobuf==5.28.1
scikit_learn==1.4.2
vertexai==1.66.0
xgboost==1.5.0
google-cloud-aiplatform
Jinja2
pyarrow==10.0.1